requirements.txt
fastapi
uvicorn
playwright
sqlalchemy
psycopg2
dotenv
alembic
fastapi-users
pandas
openpyxl
celery
jwt
python-jose
passlib
bcrypt==3.2.0
authlib==1.2.1
requests==2.31.0
python-multipart==0.0.6
aiofiles==0.23.1
httpx==0.27.0
APScheduler==3.10.4          # Advanced Python Scheduler
pytz==2023.3                 # Timezone handling
asyncio-cron==0.1.2          # Async cron support (optional)
cryptography>=41.0.0